Inline runInit()
This commit inlines runInit(), so that it is easier to repeat the
runDatabaseInitCluster() call when needed.
